# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RepoBoost API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface User {
  id: number;
  githubId: string;
  username: string;
  email?: string;
  avatarUrl?: string;
  credits: number;
  createdAt?: string;
}

export interface Repository {
  id: number;
  name: string;
  fullName: string;
  description?: string;
  language?: string;
  stars?: number;
  updatedAt?: string;
  url: string;
  private: boolean;
}

export interface RepositoryList {
  repos: Repository[];
}

export interface GenerateRequest {
  repoFullName: string;
  repoName: string;
  repoDescription?: string;
  repoUrl?: string;
  repoLanguage?: string;
}

export interface LinkedInPost {
  title: string;
  content: string;
  tone: string;
  targetAudience: string;
}

export interface GenerateResponse {
  generationId: number;
  readme: string;
  linkedinPosts: LinkedInPost[];
  creditsRemaining: number;
}

export interface Generation {
  id: number;
  repoName: string;
  repoFullName: string;
  readme: string;
  linkedinPosts: LinkedInPost[];
  createdAt: string;
}

export interface GenerationList {
  generations: Generation[];
}

export interface CreditBalance {
  credits: number;
}

export type CheckoutRequestPackageId =
  (typeof CheckoutRequestPackageId)[keyof typeof CheckoutRequestPackageId];

export const CheckoutRequestPackageId = {
  pack_10: "pack_10",
  pack_25: "pack_25",
  pack_50: "pack_50",
} as const;

export interface CheckoutRequest {
  packageId: CheckoutRequestPackageId;
}

export interface CheckoutResponse {
  checkoutUrl: string;
}

export type GithubAuthCallbackParams = {
  code: string;
  state?: string;
};