# lib/api-zod/src/generated/types/generateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RepoBoost API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateRequest {
  repoFullName: string;
  repoName: string;
  repoDescription?: string;
  repoUrl?: string;
  repoLanguage?: string;
}