# lib/api-zod/src/generated/types/generateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RepoBoost API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LinkedInPost } from "./linkedInPost";

export interface GenerateResponse {
  generationId: number;
  readme: string;
  linkedinPosts: LinkedInPost[];
  creditsRemaining: number;
}