# lib/api-zod/src/generated/types/generation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RepoBoost API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LinkedInPost } from "./linkedInPost";

export interface Generation {
  id: number;
  repoName: string;
  repoFullName: string;
  readme: string;
  linkedinPosts: LinkedInPost[];
  createdAt: Date;
}