# lib/api-zod/src/generated/types/generationList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RepoBoost API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Generation } from "./generation";

export interface GenerationList {
  generations: Generation[];
}