# lib/api-zod/src/generated/types/linkedInPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RepoBoost API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LinkedInPost {
  title: string;
  content: string;
  tone: string;
  targetAudience: string;
}