# lib/api-zod/src/generated/types/repository.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RepoBoost API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Repository {
  id: number;
  name: string;
  fullName: string;
  description?: string;
  language?: string;
  stars?: number;
  updatedAt?: Date;
  url: string;
  private: boolean;
}