# lib/api-zod/src/generated/types/repositoryList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RepoBoost API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Repository } from "./repository";

export interface RepositoryList {
  repos: Repository[];
}