# lib/db/src/schema/generations.ts
import { pgTable, text, serial, integer, timestamp, jsonb } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { usersTable } from "./users";

export const generationsTable = pgTable("generations", {
  id: serial("id").primaryKey(),
  userId: integer("user_id").notNull().references(() => usersTable.id),
  repoName: text("repo_name").notNull(),
  repoFullName: text("repo_full_name").notNull(),
  repoDescription: text("repo_description"),
  repoLanguage: text("repo_language"),
  readme: text("readme").notNull(),
  linkedinPosts: jsonb("linkedin_posts").notNull(),
  createdAt: timestamp("created_at").defaultNow().notNull(),
});

export const insertGenerationSchema = createInsertSchema(generationsTable).omit({ id: true, createdAt: true });
export type InsertGeneration = z.infer<typeof insertGenerationSchema>;
export type Generation = typeof generationsTable.$inferSelect;